chore(mrs): move deprecated resources to the 'deprecated' directory
